rootston: remove old cursor implementation
